Expose CLI for `egglog-experimental` (#510)
* Clean up existing CLI

* Factor out CLI

* Document Args struct and remove unused options

* Code motion

* Try to fix wasm build

* Remove cfgs

* Remove bin deps

* Alphebetize dependencies

* Always export FULL_VERSION

* Add back bin feature

* Add back serde and graphviz features

* Tabs to spaces in toml
